test(api): share service fixtures
